+++ compute_endpoint/globus_compute_endpoint/engines/high_throughput/zmq_pipes.py
@@ -5,7 +5,7 @@ import time
 
 import dill
 import zmq
-from globus_compute_endpoint.executors.high_throughput.messages import Message
+from globus_compute_endpoint.engines.high_throughput.messages import Message
 
 log = logging.getLogger(__name__)
